pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi-pagination>=0.15.8",
    "fastapi[standard]>=0.128.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
]
